Parse tenant->config->options into an object...
When init() of a tenant we are storing the options under the _config. However when coming from the database the options are stored as a string and later on in the file we are trying to read or write to them as if they are a JSON object. This bit of code will check if it is a string and if so attempt to parse it into an object so the methods later on will work as expected.